Delete user
Didn't know how to test this code. While I have tested deleting a user in SQL, and so I think it should work with this too. Wanted to make sure it is right, so pushing it to this patch. Can be deleted once confirmed and then pull request can be made in geartracker_backend.